Tweaks for production

+++ migrations/20221219205523_drop-route-specific-columns.js
@@ -1,8 +1,11 @@
 exports.up = async function (knex, Promise) {
-    await knex.schema.alterTable("load_test_route_specific_data", (table) => {
-        table.dropColumn("route");
-        table.dropColumn("method");
-    });
+    // Only run this in dev. In prod we will do it manually.
+    if (process.env.NODE_ENV !== "production") {
+        await knex.schema.alterTable("load_test_route_specific_data", (table) => {
+            table.dropColumn("route");
+            table.dropColumn("method");
+        });
+    }
 };
 
 exports.down = async function (knex, Promise) {
